+++ lib/minfra/cli/cli_starter.rb
@@ -150,7 +150,13 @@ module Minfra
         root = base_path.join('hiera')
         root.join('hieradata',env_path).glob('*.eyaml').sort.each do |path|
           env_name = path.basename.sub(/(\..+)/,'').to_s
-          @envs[env_name]=Env.new(hiera_root: root, hiera_env_path: env_path, name: env_name, hiera_debug_lookups: ENV['MINFRA_DEBUG_HIERA_LOOKUPS'] == 'true')
+          @envs[env_name]=Env.new(
+            hiera_root: root, 
+            hiera_env_path: env_path, 
+            name: env_name, 
+            hiera_debug_lookups: ENV['MINFRA_DEBUG_HIERA_LOOKUPS'] == 'true', 
+            backends: ENV.fetch('MINFRA_HIERA_BACKENDS','').split(',')
+          )
         end
         @env = @envs[@env_name] # set the current env
       end